Split build and release workflow
Cleanly separate the build (push on main) and release (new version
number) workflow. The build workflow should work like it did before the
release workflow was introduced. The release workflow is not supposed
to build a new image, but instead to take the latest built image and
apply the tag of the new version to it.