fix: rebuild local docker file in local development case
In order to use a local dev version of the builder, the documentation tells the user to do the following:
```
cd gardenlinux
ln -f -s ../builder/build build
```
Furthermore, the docs suggest that the builder would re-build the docker file if necessary.

However, this is currently not working because the link detects the gardenlinux/builder.dockerfile and uses it instead of the local builder.