Add new 'source manifest' output
This outputs a list of **source** packages used in the image. This is
needed for GLVD. The list does not contain version numbers, and is
de-duplicated, so if one source package is used to create multiple
binary packages in the image, it will appear only once.

See https://github.com/gardenlinux/glvd/issues/189 for more details